model refactor
