+++ Halcyon.Web/Models/Account/ForgotPasswordRequest.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using Halcyon.Web.Filters;
+using Halcyon.Web.Filters.Validation;
 
 namespace Halcyon.Web.Models.Account
 {
