+++ Halcyon.Web/Models/User/CreateUserRequest.cs
@@ -1,5 +1,5 @@
 ﻿using Halcyon.Web.Data;
-using Halcyon.Web.Filters;
+using Halcyon.Web.Filters.Validation;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
