+++ Halcyon.Web/Services/Jwt/IJwtService.cs
@@ -4,6 +4,6 @@ namespace Halcyon.Web.Services.Jwt
 {
     public interface IJwtService
     {
-        public TokenResult GenerateToken(User user);
+        public Token CreateToken(User user);
     }
 }
